fix: Added missing fields on DTOs

+++ STAGGI-Budget-API/Repositories/Interfaces/IBudgetRepository.cs
@@ -7,7 +7,7 @@ namespace STAGGI_Budget_API.Repositories.Interfaces
     {
         IEnumerable<Budget> GetAll();
         IEnumerable<Budget> GetAllByEmail(string email);
-        public Budget GetById(long id);
+        public Budget GetById(int id);
         void Save(Budget budget);
         
         void Delete(Budget budget);
